+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # v0.0.3
 
+* Add implicit submission field inferring feature
+
 # v0.0.2
 
 * add `datetime_field` DSL method to manage Rails `datetime_select` HTML